requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
-e git+ssh://git@github.com/Belval/ebird.git@9fda290f6288c2979b61bb1afb09ac49b5add5f8#egg=ebird
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.1
Markdown==3.3.6
numpy==1.21.5
oauthlib==3.2.0
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
torch==1.10.2
torchvision==0.11.3
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
